LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/after-transition-reject-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: undefined

Harness Error (FAIL), message = Unhandled rejection

PASS scroll: after-transition should not scroll when the intercept() handler rejects
